query 관련까지
